# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TeacherAttendanceSettingsDeductionType = typeof TeacherAttendanceSettingsDeductionType[keyof typeof TeacherAttendanceSettingsDeductionType];


export const TeacherAttendanceSettingsDeductionType = {
  daily_rate: 'daily_rate',
  fixed: 'fixed',
} as const;

export interface TeacherAttendanceSettings {
  /** @nullable */
  schoolLatitude?: number | null;
  /** @nullable */
  schoolLongitude?: number | null;
  radiusMeters?: number;
  checkInStart?: string;
  checkInEnd?: string;
  checkOutStart?: string;
  checkOutEnd?: string;
  requireFaceVerification?: boolean;
  workingDaysPerMonth?: number;
  lateGraceMinutes?: number;
  lateDeductionAmount?: number;
  deductionType?: TeacherAttendanceSettingsDeductionType;
}

export type TeacherAttendanceSettingsUpdate = TeacherAttendanceSettings & {
  adminId: string;
};

export type TeacherAttendanceRecordStatus = typeof TeacherAttendanceRecordStatus[keyof typeof TeacherAttendanceRecordStatus];


export const TeacherAttendanceRecordStatus = {
  present: 'present',
  late: 'late',
  absent: 'absent',
  leave: 'leave',
} as const;

export interface TeacherAttendanceRecord {
  id?: string;
  teacherId?: string;
  teacherName?: string;
  date?: string;
  status?: TeacherAttendanceRecordStatus;
  /** @nullable */
  checkInAt?: string | null;
  /** @nullable */
  checkOutAt?: string | null;
  /** @nullable */
  checkInLatitude?: number | null;
  /** @nullable */
  checkInLongitude?: number | null;
  /** @nullable */
  checkOutLatitude?: number | null;
  /** @nullable */
  checkOutLongitude?: number | null;
  /** @nullable */
  distanceFromSchool?: number | null;
  faceVerified?: string;
  /** @nullable */
  faceVerificationMethod?: string | null;
}

export interface TeacherCheckIn {
  teacherId: string;
  teacherName: string;
  date?: string;
  latitude: number;
  longitude: number;
  faceVerified: boolean;
  faceVerificationMethod?: string;
}

export interface TeacherCheckOut {
  teacherId: string;
  latitude: number;
  longitude: number;
}

export type TeacherLeaveStatus = typeof TeacherLeaveStatus[keyof typeof TeacherLeaveStatus];


export const TeacherLeaveStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface TeacherLeave {
  id?: string;
  teacherId?: string;
  teacherName?: string;
  startDate?: string;
  endDate?: string;
  reason?: string;
  status?: TeacherLeaveStatus;
  /** @nullable */
  adminNote?: string | null;
}

export interface TeacherLeaveCreate {
  teacherId: string;
  teacherName: string;
  startDate: string;
  endDate: string;
  reason: string;
}

export interface AdminAction {
  adminId: string;
  adminNote?: string;
}

export interface TeacherHoliday {
  id?: string;
  date?: string;
  name?: string;
}

export interface TeacherHolidayCreate {
  adminId: string;
  date: string;
  name: string;
}

export interface PayrollCalculate {
  adminId: string;
  month: string;
  year: number;
}

export type PayrollReportResultItem = { [key: string]: unknown };

export interface PayrollReport {
  month?: string;
  year?: number;
  workingDays?: number;
  result?: PayrollReportResultItem[];
}

export type TeacherIdParameter = string;

export type MonthParameter = string;

export type StatusParameter = string;

export type AdminIdParameter = string;

export type ListTeacherAttendanceParams = {
teacherId?: TeacherIdParameter;
month?: MonthParameter;
};

export type ListTeacherLeavesParams = {
teacherId?: TeacherIdParameter;
status?: StatusParameter;
};

export type DeleteTeacherHolidayParams = {
adminId: AdminIdParameter;
};