# lib/api-zod/src/generated/types/teacherAttendanceRecord.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TeacherAttendanceRecordStatus } from './teacherAttendanceRecordStatus';

export interface TeacherAttendanceRecord {
  id?: string;
  teacherId?: string;
  teacherName?: string;
  date?: string;
  status?: TeacherAttendanceRecordStatus;
  /** @nullable */
  checkInAt?: Date | null;
  /** @nullable */
  checkOutAt?: Date | null;
  /** @nullable */
  checkInLatitude?: number | null;
  /** @nullable */
  checkInLongitude?: number | null;
  /** @nullable */
  checkOutLatitude?: number | null;
  /** @nullable */
  checkOutLongitude?: number | null;
  /** @nullable */
  distanceFromSchool?: number | null;
  faceVerified?: string;
  /** @nullable */
  faceVerificationMethod?: string | null;
}