# lib/api-zod/src/generated/types/teacherAttendanceSettings.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TeacherAttendanceSettingsDeductionType } from './teacherAttendanceSettingsDeductionType';

export interface TeacherAttendanceSettings {
  /** @nullable */
  schoolLatitude?: number | null;
  /** @nullable */
  schoolLongitude?: number | null;
  radiusMeters?: number;
  checkInStart?: string;
  checkInEnd?: string;
  checkOutStart?: string;
  checkOutEnd?: string;
  requireFaceVerification?: boolean;
  workingDaysPerMonth?: number;
  lateGraceMinutes?: number;
  lateDeductionAmount?: number;
  deductionType?: TeacherAttendanceSettingsDeductionType;
}