# lib/api-zod/src/generated/types/teacherCheckIn.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TeacherCheckIn {
  teacherId: string;
  teacherName: string;
  date?: string;
  latitude: number;
  longitude: number;
  faceVerified: boolean;
  faceVerificationMethod?: string;
}